Perfect hash dictionary performance * Refactor perfect hash table * Added dictionary perf for perfect hash table

+++ library/include/utils/hash_string.h
@@ -2,4 +2,4 @@
 
 #include <string>
 
-int hashString(std::string str);
+size_t hashString(std::string str);
